move config editor to lazy-loaded ./Tabulator/ConfigEditor
